big_data_analysis/assignment5/pyproject.toml
[tool.poetry]
name = "assignment5"
version = "0.1.0"
description = ""
authors = ["Aleksandr J. Smoliakov <aleksandr.smoliakov@mif.stud.vu.lt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pyspark = "^3.5.5"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
numpy = "^2.2.6"
matplotlib = "^3.10.3"
wordcloud = "^1.9.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
